genBileDuct/requirements.txt
einops==0.4.1
h5py==3.3.0
h5py==2.10.0
imageio==2.9.0
imgaug==0.4.0
matplotlib==3.4.2
matplotlib==3.3.4
numpy==1.20.2
Pillow==8.2.0
Pillow==9.5.0
Requests==2.31.0
rich==13.3.5
scikit_image==0.18.1
scipy==1.6.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
util==0.0.1
